e2e: Refactoring namespace-related test functions into a new file - namespace.go
